+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Add `SymbolEntryIndex.getSymbolEntries()` method that returns all the entries in the order they appear
+  in the input file.
+
 ### [v0.02](https://github.com/realityforge/gwt-symbolmap/tree/v0.02) (2018-03-15)
 [Full Changelog](https://github.com/realityforge/gwt-symbolmap/compare/v0.01...v0.02)